inlining - minor refactor
Introduce dwarf helpers to simplify some of the structure
